+++ Makefile
@@ -54,7 +54,7 @@ lint: ## check style with flake8
 	flake8 xclim tests
 
 test: ## run tests quickly with the default Python
-	py.test
+	pytest xclim/testing/tests
 
 test-all: ## run tests on every Python version with tox
 	tox
