+++ setup.cfg
@@ -50,7 +50,7 @@ collect_ignore = ["setup.py"]
 addopts = --verbose
 filterwarnings =
 	ignore::UserWarning
-testpaths = tests tests/test_sdba
+testpaths = xclim/testing/tests xclim/testing/tests/test_sdba
 usefixtures = xdoctest_namespace
 doctest_optionflags = NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL NUMBER ELLIPSIS
