--- tests/data.py
@@ -1,6 +1,0 @@
-import os
-from pathlib import Path
-
-from xclim.testing import TestDataSet
-
-TD = Path(os.path.dirname(__file__)) / "testdata"
