+++ xclim/testing/tests/__init__.py
@@ -1,0 +1,2 @@
+from .data import TD
+from .test_sdba import *
