+++ xclim/testing/tests/data.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+__all__ = ["TD"]
+
+TD = Path(__file__).parent / "data"
