+++ xclim/testing/tests/data/small_geojson.json
