+++ xclim/testing/tests/test_analog.py
