+++ xclim/testing/tests/test_atmos.py
