+++ xclim/testing/tests/test_calendar.py
