+++ xclim/testing/tests/test_checks.py
