+++ xclim/testing/tests/test_ensembles.py
