+++ xclim/testing/tests/test_formatting.py
