+++ xclim/testing/tests/test_fwi.py
