+++ xclim/testing/tests/test_generic.py
