+++ xclim/testing/tests/test_indicators.py
