+++ xclim/testing/tests/test_locales.py
