+++ xclim/testing/tests/test_modules.py
