+++ xclim/testing/tests/test_options.py
