+++ xclim/testing/tests/test_precip.py
