+++ xclim/testing/tests/test_sdba/__init__.py
@@ -1,0 +1,2 @@
+from .diagnostics import *
+from .utils import *
