+++ xclim/testing/tests/test_seaice.py
