+++ xclim/testing/tests/test_stats.py
