+++ xclim/testing/tests/test_temperature.py
